Move tests to integration
